feat: upgrade to go1.19
