+++ .dockerignore
@@ -7,4 +7,3 @@ action.yml
 Dockerfile
 LICENSE
 README.md
-slim.dockerfile
