+++ go.mod
@@ -1,12 +1,11 @@
 module github.com/maolonglong/actions-starcharts
 
-go 1.17
+go 1.19
 
 require (
 	github.com/google/go-github/v39 v39.2.0
-	github.com/sethvargo/go-githubactions v0.5.3
+	github.com/sethvargo/go-githubactions v1.0.0
 	github.com/shurcooL/githubv4 v0.0.0-20210922025249-6831e00d857f
-	github.com/spf13/cast v1.4.1
 	github.com/wcharczuk/go-chart v2.0.1+incompatible
 	golang.org/x/oauth2 v0.0.0-20210819190943-2bc19b11175f
 )
@@ -16,8 +15,8 @@ require (
 	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
 	github.com/golang/protobuf v1.5.2 // indirect
 	github.com/google/go-querystring v1.1.0 // indirect
+	github.com/sethvargo/go-envconfig v0.6.0 // indirect
 	github.com/shurcooL/graphql v0.0.0-20200928012149-18c5c3165e3a // indirect
-	github.com/stretchr/testify v1.5.1 // indirect
 	golang.org/x/crypto v0.0.0-20210817164053-32db794688a5 // indirect
 	golang.org/x/image v0.0.0-20190802002840-cff245a6509b // indirect
 	golang.org/x/net v0.0.0-20210929193557-e81a3d93ecf6 // indirect
